kafka: fix another data race in tests
This is a bit unfortunate, but if we try to properly clean stuff, we
get a race because Close() spawns a goroutine to close brokers and
this will use the logger at some point, so calling kafka.New() will
trigger a race.

We prefer to keep logging working since it makes it easier to debug
tests.