snmp: lower even more the default value fo coalescing
If we assume a MTU of 1500, we may run into trouble with long
descriptions. Assume 50 bytes for description, 30 bytes for interface
name, plus 30 bytes for OIDs, it means we can only put 13 results in a
standard MTU.

At some point, we may try to be smarter and have per-sampler values?