orchestrator/clickhouse: create consolidated tables for flows
The idea is to not query the flows table unless absolutely necessary.
It would have been nice to not have this Date field, but rebuilding
the table is costly, we'll do that later when the table is smaller. We
will also need to use a small PARTITION BY.

Also remove some migrations not needed anymore.