inlet/flow: version flow message names as well
ClickHouse do not like to have several proto files with the same
message name. We also put a version in the message name to solve this.
We use a hack to have this invisble from Go (struct FlowMessage
instead of struct FlowMessagevX). This works because the descriptor is
already encoded (and contains FlowMessagevX). Moreover, FlowMessage
"v1" is still FlowMessage as it will be in use when users will
upgrade. FlowMessagev0 was never used in a released version, so we
don't care.