docker-compose: add a service to clean conntrack entries
When the inlet container starts, there is a race condition between the
moment Docker sets up a competing listening socket on the host to
ensure the port is free and the moment it sets up the NAT rules. If
packets are received during this interval, a conntrack entry is
created and the NAT rule is ignored. These entries need to be flushed.
Docker tries to do that but is not very successful. The service watch
docker events and flush the needed conntrack entries when the
container starts.