orchestrator/clickhouse: fix `SrcCountry`/`DstCountry` columns
In aggregated tables, these columns were missing from the ORDER BY
clause. This means they were set to some random values. This is not
possible to fix that after their creation (see #60 for a tentative),
therefore, we have to drop and recreate the columns. This only affects
aggregated tables, not the main table, but nonetheless, unless you
look at the last hour, the data is lost.