conntrack-fixer: don't listen to 8080
Otherwise, we take port from Traefik. Also, this makes the healthcheck
ineffective. Maybe there is a way to escape from the network namespace
we are put in just for conntrack, but I don't see exactly how without
doing complicated stuff (like using Docker to spawn a process inside
the host namespace and communicate with it, or even steal the
namespace from it, this seems overkill).