inlet/geoip: use low-level maxminddb-golang to query GeoIP database
The higher-level geoip2-golang will decode far more records when
requesting for a country. This takes time and since we do that for
each packet, this is costly. Half of the CPU time of the inlet is
spent looking up into this database. This should improve performance
quite a bit.