inlet/core: replace LRU cache by TTL cache
It is easy for a user to configure a cache too small for the classifier.
Instead, use a time-based cache. Moreover, the cache (from @arp242) is
using generics, so it's likely adding a slight performance improvement
over the previous method (no need to use interfaces and no need to
serialize the key to string).