inlet/flow: handle correctly interfaces with high index for sFlow
There are two kind of flow samples in sFlow: regular ones and expanded
ones. For the regular ones, the first two bits of the interface are used
for meta data. For the expanded ones, there is a dedicated 32-bit
integer for that (however, no documentation of what it means).
Unfortunately, the information is lost after the decoding abstraction
done by GoFlow2. Therefore, we take note of the kind of samples we get
to do the decoding.

However, we take a shortcut and assume all flows in a packet are either
the regular ones or the expanded ones. This is less error-prone than
trying to do more complex stuff. Let's see if this is really the case or
not.

Fix #370