inlet/core: convert classifier cache to internal cache helper
This also reduce the amount of `time.Now()` to one per flow. Far lower
than it previously was. `time.Now()` should usually be quick, but on
some VM, it can be slow (100 times) due to a missing or dysfunctional
TSC.

We also modify the SNMP component to not rely on clock either (it must be
externally provided). Internally, it can still use the system clock (not
on the hot path).