orchestrator/clickhouse: build networks.csv on the fly
With the IPinfo GeoIP database, we have 1.3GB of memory used by
ClickHouse and 3.3GB of memory used inside the orchestrator, plus 3.3GB
of memory used during updates. This is quite a step up compared to
almost no memory used, except caching when GeoIP was done in the inlet.

Instead, we build the data on the fly. It is requested by ClickHouse on
each database update (configurable for network sources, more seldomly
for GeoIP database) and every hour with the default configuration.
During this update, it will use 1.5GB of memory, so the max usage could
be 1.3+1.5GB = 3.8GB against 1.3+3.3+3.3GB = 7.9GB (the double).

It takes about 30s to build the dictionary on the fly (on Intel(R)
Xeon(R) W-2295 CPU @ 3.00GHz).