build: remove functional tests for GitLab CI
It was working, but this is a bit of work to keep it working. Now that
we use docker compose with GitHub CI, GitLab CI starts to diverge. As
GitLab setup often run dind as a runner, I don't know if it is as easy
to run docker compose in GitLab and I don't have time to test it.

If someone is interested in this support, there are two possibilities:
use act to run GitHub CI from GitLab or also adapt docker compose
workflow to GitLab CI (but it should work with docker-based runners).