orchestrator: do not use AS names for tenants
This was introduced with #1059, but I think this was a mistake. Notably,
it enables erasing the tenants provided by the user.

It also opens the question whetever to have network sources or static
sources override more specific entries or not. This is currently not the
case, but then, if a more specific GeoIP entry appears, it may require
to add a more specific entry if overriding is needed.

This could also be configurable.