chore: use time.NewTimer() instead of time.After()
Prefer use of time.NewTimer() when there is a risk of accumulating
timers in a loop. This enables the use of t.Stop() to avoid leaking too
many timers.

For tests, we don't need to do that. For places where the alternative to
the timer is just the app dying, we don't need to do that either.

In Go 1.23, it won't make a difference.