Currently, for discovering interfaces' name, description and speed, the gNMI metadata provider assumes that all entries returned by the device always contain a single key.
However, this is not always the case. For example, on Nokia SR OS, sampling is enabled on logical interfaces that are defined inside router instances or VPRN services, rather than physical ports.

To give an example:
```
$ gnmic get -u akvorado ... \
    --path 'state/service/vprn/interface/system-if-index'
[
  {
...
    "updates": [
      {
        "Path": "state/service/vprn[service-name=WAN_VRF]/interface[interface-name=primary_transit]/system-if-index",
        "values": {
          "state/service/vprn/interface/system-if-index": 264
        }
      },
...
```

Adding this path to if-index-paths, with interface-name as a if-name-keys, would not allow primary_transit to be successfully discovered as the name for interface index 264.

The proposed change aims at adding support for this type of situation.