inlet/metadata: prefer ifAlias over ifDescr for interface description
ifDescr is more likely to contain the interface name for compatibility
with ancien stuff (part of the original ifTable). The heuristics about
it being different than ifName is not 100% correct as some IOS XE device
may use the short name for ifDescr and the long name for ifName.

ifAlias contains the description for IOS XR and JunOS.