docker: switch to Apache Kafka
This change is not transparent because we can't easily guess the current
cluster ID, like this is done in Bitnami images. Also note that there is
a tentative to make an official Docker image, which should increase the
quality of the image.

- https://cwiki.apache.org/confluence/display/KAFKA/KIP-1028%3A+Docker+Official+Image+for+Apache+Kafka
- https://issues.apache.org/jira/browse/KAFKA-16978

This is part of the effort to not rely on Bitnami images anymore.