docker: make Alloy configuration use only Docker labels
This is a bit like Traefik. We set metrics.port on each container we
want to scrape metrics from (and optionally metrics.path).

Semi-related, but we also rely on exposed port for Traefik and we override
it for all containers to be sure we select the right one. This is less
error prone as we need at least one exposed port and some containers may
or may not have one. Just always set an exposed port if we have metrics
or traefik rules.