outlet/kafka: expose consumer lag as a prometheus metric
Monitoring consumer lag is useful to troubleshoot performance/scaling
issues. It can currenctly be seen through kafka-ui, but a proper metric
is more practical.

Unfortunately, JMX metrics on the broker don't expose this. It seems
that people usually resort to monitoring from the consumer side, or
through other external exporters like Burrow or kafka_exporter.

franz-go/kadm provides a function to compute the consumer lag so let's
do it from the consumer side (the outlet)