build: rearrange CI pipeline
Test Go and JS first, then more complex builds. This delay a bit "Test
on Linux", but maybe the cache will make for it. This way, we ensure the
cache is "complete" for Linux and e2e. Mac would not need to be delayed,
but as it is shorter, it does not matter much.