build: update Go toolchain to the latest patch version
This is a bit a continuation of 94fb1c64. We don't want to have a
specific Go version in "go" directive, so we keep a language version.
People not wanting to use an up-to-date toolchain can still use a not
up-to-date one. But for the others, we want to be sure we get the latest
toolchain for the language version and this is not something automatic
as without a toolchain directive, go will pull the first version (eg.
1.25.0).

Also add a check that "go" directive is always a language version.