inlet/flow: ensure we request 64-bit timestamps from the kernel
This requires Linux 5.0+. Below, we would just get no timestamp. This is
more correct this way, even if most people would run that on 64-bit
Linux and already get 64-bit timestamp.

We also don't use the nanosecond part as it is "long long" and should be
virtually 64-bit on all archs, this is not totally correct.