console: don't use function call in text/template
This requires usage of `reflect.MethodByName`, which prevents DCE. This
does not really matter as DCE is not smart enough to detect we don't use
it, but the day we have an alternative not supporting functions call, we
would be ready. Or maybe there is an alternative. Or we could use
`strings.ReplaceAll()` instead.